requirements.txt
kaggle
inference
supervision
Pillow
roboflow

# für das Laden der Informationen aus der .env-Datei
python-dotenv
